+++ src/main/java/com/ll/medium/comment/CommentController.java
@@ -1,21 +1,20 @@
 package com.ll.medium.comment;
 
-import com.ll.medium.article.Article;
+import com.ll.medium.article.entity.Article;
 import com.ll.medium.article.ArticleService;
-import com.ll.medium.member.Member;
-import com.ll.medium.member.MemberService;
-import jakarta.persistence.criteria.CriteriaBuilder;
+import com.ll.medium.comment.entity.Comment;
+import com.ll.medium.comment.entity.CommentForm;
+import com.ll.medium.member.entity.Member;
+import com.ll.medium.member.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
-import org.testng.TestNGAntTask;
 
 import java.security.Principal;
 import java.util.List;
