+++ src/main/java/com/ll/medium/comment/CommentRepository.java
@@ -1,6 +1,7 @@
 package com.ll.medium.comment;
 
-import com.ll.medium.article.Article;
+import com.ll.medium.article.entity.Article;
+import com.ll.medium.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
