+++ src/main/java/com/ll/medium/member/entity/MemberCreateForm.java
@@ -1,4 +1,4 @@
-package com.ll.medium.member;
+package com.ll.medium.member.entity;
 
 
 import jakarta.validation.constraints.NotEmpty;
