+++ src/main/java/com/ll/medium/member/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package com.ll.medium.member;
+package com.ll.medium.member.entity;
 
 import lombok.Getter;
